Unify font type part 2
- Also change build target from dynamic to static library
- Remove dynamic link library entry point

+++ src/controls/combobox.cc
@@ -76,7 +76,7 @@ void zgui::multi_combobox(const char* id, std::vector<multi_select_item> items) 
 	const int control_width = 100;
 	const int control_height = 20;
 
-	const int font = utils::misc::pop_font();
+	const unsigned long font = utils::misc::pop_font();
 
 	const vec2 cursor_pos = utils::misc::pop_cursor_pos();
 	vec2 draw_pos{ context.window.position.x + cursor_pos.x + 14, context.window.position.y + cursor_pos.y };
